+++ src/components/SamplePlayer.svelte
@@ -26,10 +26,11 @@
     velocityCurveHigh,
   } from "../stores";
 
+  export let webMidi;
+
   let tempoMap;
   let pedalingMap;
   let notesMap;
-  let midiOuts = [];
 
   const SOFT_PEDAL = 67;
   const SUSTAIN_PEDAL = 64;
@@ -67,12 +68,6 @@
 
   const pianoReady = piano.load();
 
-  const sendMidiMsg = (msg) => {
-    for (let i = 0; i < midiOuts.length; i++) {
-      midiOuts[i].send(msg);
-    }
-  };
-
   const getTempoAtTick = (tick) => {
     if (!tempoMap || !$useMidiTempoEventsOnOff) return DEFAULT_TEMPO;
     let tempo;
@@ -91,8 +86,8 @@
       return;
     }
     onOff ? piano.pedalDown() : piano.pedalUp();
-    if (midiOuts.length && !fromMidi) {
-      sendMidiMsg([MIDI_CONTROL, MIDI_SUSTAIN, (onOff ? 1 : 0) * 127]);
+    if (!fromMidi) {
+      webMidi?.sendMidiMsg([MIDI_CONTROL, MIDI_SUSTAIN, (onOff ? 1 : 0) * 127]);
     }
   };
 
@@ -101,8 +96,8 @@
       $softOnOff = onOff;
       return;
     }
-    if (midiOuts.length && !fromMidi) {
-      sendMidiMsg([MIDI_CONTROL, MIDI_SOFT, (onOff ? 1 : 0) * 127]);
+    if (!fromMidi) {
+      webMidi?.sendMidiMsg([MIDI_CONTROL, MIDI_SOFT, (onOff ? 1 : 0) * 127]);
     }
   };
 
@@ -212,8 +207,8 @@
         velocity: Math.min(modifiedVelocity, 1),
       });
     }
-    if (midiOuts.length && !fromMidi) {
-      sendMidiMsg([
+    if (!fromMidi) {
+      webMidi?.sendMidiMsg([
         MIDI_NOTE_ON,
         noteNumber,
         parseInt(modifiedVelocity * 127, 10),
@@ -223,8 +218,8 @@
 
   const stopNote = (noteNumber, fromMidi) => {
     piano.keyUp({ midi: noteNumber });
-    if (midiOuts.length && !fromMidi) {
-      sendMidiMsg([MIDI_NOTE_OFF, noteNumber, 0]);
+    if (!fromMidi) {
+      webMidi?.sendMidiMsg([MIDI_NOTE_OFF, noteNumber, 0]);
     }
   };
 
@@ -307,53 +302,6 @@
     return _notesMap;
   };
 
-  const registerMidiInputs = (midi) => {
-    // Respond to input from attached MIDI controllers
-    Array.from(midi.inputs).forEach((input) => {
-      if (input[1].onmidimessage !== null) return;
-      input[1].onmidimessage = (msg) => {
-        if (msg.data.length > 1) {
-          if (msg.data[0] == MIDI_CONTROL) {
-            if (msg.data[1] == MIDI_SUSTAIN) {
-              toggleSustain(!!parseInt(msg.data[2], 10), true);
-            } else if (msg.data[1] == MIDI_SOFT) {
-              toggleSoft(!!parseInt(msg.data[2], 10), true);
-            }
-          } else if (msg.data[0] == MIDI_NOTE_ON) {
-            if (msg.data[2] === 0) {
-              stopNote(msg.data[1], true);
-              activeNotes.delete(msg.data[1]);
-            } else {
-              startNote(
-                msg.data[1],
-                parseInt((parseFloat(msg.data[2]) / 127.0) * 100.0, 10),
-                true,
-              );
-              activeNotes.add(msg.data[1]);
-            }
-          } else if (msg.data[0] == MIDI_NOTE_OFF) {
-            stopNote(msg.data[1], true);
-            activeNotes.delete(msg.data[1]);
-          }
-        }
-      };
-    });
-  };
-
-  const registerMidiDevices = () => {
-    if (navigator.requestMIDIAccess) {
-      navigator.requestMIDIAccess().then((midi) => {
-        registerMidiInputs(midi);
-        midiOuts = Array.from(midi.outputs).map((output) => output[1]);
-        // This handles when new devices are connected (or disconnected)
-        midi.onstatechange = () => {
-          registerMidiInputs(midi);
-          midiOuts = Array.from(midi.outputs).map((output) => output[1]);
-        };
-      });
-    }
-  };
-
   midiSamplePlayer.on("fileLoaded", () => {
     const decodeHtmlEntities = (string) =>
       string
@@ -376,8 +324,6 @@
           ),
       ),
     );
-
-    registerMidiDevices();
 
     tempoMap = buildTempoMap(metadataTrack);
 
@@ -436,5 +382,8 @@
     pausePlayback,
     startPlayback,
     resetPlayback,
+    activeNotes,
+    toggleSustain,
+    toggleSoft,
   };
 </script>
